fix: correctly handle arithmetics of the same precedence
In order to handle arithmetics correctly, operators of the same
precedence should not be handled in a recursive way. The easiest way was
to just convert the parameters in a list, when they are from the same
precedence and let the javascript arithmetics figure it out. This
StackOverflow post was really helpful:
http://stackoverflow.com/questions/19390084/parsing-complete-mathematical-expressions-with-peg-js#answer-30798758

fixes #108